# application/services/custom_services/contractors/src/swagger/swagger.yaml
---
openapi: 3.0.0
info:
   version: 1.3.0
   title: cheiconstru open api specification
   description: contractors API
   termsOfService: http://api_url/terms/
   contact:
      name: ContactName
      email: hello@wolox.co
      url: https://www.wolox.com.ar/
   license:
      name: Apache 2.0
      url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
   - url: http://localhost:8017
     description: contractors
tags:
- name: contractors
paths:
   "/contractors/{id}":
      delete:
         tags:
         -  contractors
         description: GpDelete
         parameters:
         -  in: "path"
            description: GpDelete
            name: id
            required: true
            schema:
               type: "string"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      get:
         tags:
         -  contractors
         description: GpGetEntityById
         parameters:
         -  in: "path"
            description: GpGetEntityById
            name: id
            required: true
            schema:
               type: "string"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/contractors/get/search":
      get:
         tags:
         -  contractors
         description: GpSearch
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/contractors/get/update":
      put:
         tags:
         -  contractors
         description: GpSearchForUpdate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/contractors"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/contractors":
      put:
         tags:
         -  contractors
         description: GpUpdate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/contractors"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      get:
         tags:
         -  contractors
         description: GpGetAllValues
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      post:
         tags:
         -  contractors
         description: GpCreate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/contractors"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/contractors/userid/created_by":
      get:
         tags:
         -  contractors
         description: GpGetNounCreatedBy
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
components:
   schemas:
      contractors:
         type: object
         properties:
            contractor_name:
               type: "string"
               description: Description
            company_name:
               type: "string"
               description: Description
            mobile:
               type: "string"
               description: Description
            email:
               type: "string"
               description: Description
            contractorstatus:
               type: "string"
               description: Description
            zip_code:
               type: "string"
               description: Description
            proof_of_insurance:
               type: "string"
               description: Description
            proof_of_address:
               type: "string"
               description: Description
            address_line_1:
               type: "string"
               description: Description
            address_line_2:
               type: "string"
               description: Description
            city:
               type: "string"
               description: Description